web_app/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
turboloader>=0.8.0
torch>=2.0.0
pillow>=10.0.0
